Create `matrixui` module and remove dependency to `coil` in the matrix module. Move `MatrixUser` there.
