+++ app/src/main/java/io/element/android/x/di/AppBindings.kt
@@ -18,11 +18,13 @@ package io.element.android.x.di
 
 import com.squareup.anvil.annotations.ContributesTo
 import io.element.android.x.matrix.Matrix
+import io.element.android.x.matrix.ui.MatrixUi
 import kotlinx.coroutines.CoroutineScope
 
 @ContributesTo(AppScope::class)
 interface AppBindings {
     fun coroutineScope(): CoroutineScope
     fun matrix(): Matrix
+    fun matrixUi(): MatrixUi
     fun sessionComponentsOwner(): SessionComponentsOwner
 }
