+++ settings.gradle.kts
@@ -37,6 +37,7 @@ include(":app")
 include(":libraries:core")
 include(":libraries:rustsdk")
 include(":libraries:matrix")
+include(":libraries:matrixui")
 include(":libraries:textcomposer")
 include(":libraries:elementresources")
 include(":features:onboarding")
